Restructured; planning log files
It would be cool to use futures here too. There will be a bunch of
buffers for different channels (including private conversations) that
will be updated with extremely unpredictable frequency (once ever to
dozens of times per second). Instead of maintaining a bunch of
BufWriters that are always open, or always closing a bufwriter after
touching it, it would be interesting to use a timer (tokio-timer) to
close and files that haven't been touched in a certain tweakable amount
of time. How troublesome is that? We'll see